I never found :nu useful, but :rnu is
The status line already has the current line number and finding the
the number for other lines is not as usefule as the relative number.

In the newest version of vim, the :set nu rnu is nice, freeing up some
space in the status line for other useful stuff at a later date.