Fixed XDG paths
The paths were written with a naive understanding of how environment
variable substitution works in vim.  This issue was not apparent
earlier, because I used to explicitly set XDG variables in my
environment.

Previously this would create a directory called '~' in the working
directory.  This caused issues if you wanted to delete it from the
working directory like this: `rm -r ~`

This would instead delete your home directory.